requirements.txt
supabase==2.24.0
supabase-auth==2.24.0
supabase-functions==2.24.0
streamlit==1.51.0
fpdf==1.7.2
pandas==2.1.0
python-dotenv==1.0.0
